third_party/blink/web_tests/external/wpt/css/css-box/margin-trim/computed-margin-values/flexbox-column-inline-end-expected.txt

This is a testharness.js-based test.
[FAIL] flexbox > item 1
  assert_equals: \n<item data-expected-margin-right="0" data-offset-x="8"></item>\nmargin-right expected "0" but got "10"
[FAIL] flexbox > item 2
  assert_equals: \n<item data-expected-margin-right="0" data-offset-x="8"></item>\nmargin-right expected "0" but got "-10"
[FAIL] flexbox > item 3
  assert_equals: \n<item data-expected-margin-right="0" data-offset-x="8"></item>\nmargin-right expected "0" but got "18"
Harness: the test ran to completion.
